# Improve ansi Writer's Write performance.