OLD:
```hs
BenchmarkWriter_Write-8           682036              1545 ns/op             320 B/op         25 allocs/op
BenchmarkWriter_Write-8           789792              1521 ns/op             320 B/op         25 allocs/op
BenchmarkWriter_Write-8           725499              1547 ns/op             320 B/op         25 allocs/op
BenchmarkWriter_Write-8           765880              1542 ns/op             320 B/op         25 allocs/op
```

NEW:
```hs
BenchmarkWriter_Write-8          4759159               247 ns/op               0 B/op          0 allocs/op
BenchmarkWriter_Write-8          4847602               254 ns/op               0 B/op          0 allocs/op
BenchmarkWriter_Write-8          5125417               244 ns/op               0 B/op          0 allocs/op
BenchmarkWriter_Write-8          4870538               245 ns/op               0 B/op          0 allocs/op
```