@kiyonlin 2020-10-19
NEW2:
```hs
BenchmarkWriter_Write-8          4759159               247 ns/op               0 B/op          0 allocs/op
BenchmarkWriter_Write-8          4847602               254 ns/op               0 B/op          0 allocs/op
BenchmarkWriter_Write-8          5125417               244 ns/op               0 B/op          0 allocs/op
BenchmarkWriter_Write-8          4870538               245 ns/op               0 B/op          0 allocs/op
```