@muesli 2020-10-19
Don't worry about the decreased coverage. I just changed the coveralls settings to not flag such tiny decreases as a failure.